tux3: Update slab api for v4.17

+++ fs/tux3/user/libklib/slab.c
@@ -21,9 +21,9 @@ void kfree(const void *p)
 		free((void *)p);
 }
 
-struct kmem_cache *kmem_cache_create(const char *name, size_t size,
-				     size_t align, slab_flags_t flags,
-				     void (*ctor)(void *))
+struct kmem_cache *kmem_cache_create(const char *name, unsigned int size,
+			unsigned int align, slab_flags_t flags,
+ 			void (*ctor)(void *))
 {
 	struct kmem_cache *cachep;
 
@@ -39,8 +39,9 @@ struct kmem_cache *kmem_cache_create(const char *name, size_t size,
 }
 
 struct kmem_cache *kmem_cache_create_usercopy(const char *name,
-			size_t size, size_t align, slab_flags_t flags,
-			size_t useroffset, size_t usersize,
+			unsigned int size, unsigned int align,
+			slab_flags_t flags,
+			unsigned int useroffset, unsigned int usersize,
 			void (*ctor)(void *))
 {
 	return kmem_cache_create(name, size, align, flags, ctor);
